+++ R/helper.R
@@ -542,8 +542,6 @@ deleteHelpers <- function() {
 #'
 emergencyBreak <- function(the.function.call) { # nocov start
   appendTo("Report", "\nA fatal exception occurred, stopping the process!\nFound a bug? Report it here: https://github.com/hugomflavio/actel/issues\n\n-------------------")
-  logname <- paste(gsub(":", ".", sub(" ", ".", as.character(Sys.time()))), "actel.log-ERROR.txt", sep = ".")
-
   if (file.exists(paste0(tempdir(), "/temp_comments.txt")))
     appendTo("Report", paste0("User comments:\n-------------------\n", gsub("\t", ": ", gsub("\r", "", readr::read_file(paste0(tempdir(), "/temp_comments.txt")))), "-------------------")) # nocov
 
@@ -554,9 +552,12 @@ emergencyBreak <- function(the.function.call) { # nocov start
 
   message("\nM: The analysis errored. You can recover latest the job log (including your comments and decisions) by running recoverLog().")
   
-  file.rename(paste0(tempdir(), "/temp_log.txt"), paste0(tempdir(), "/latest_actel_error_log.txt"))
-
-  deleteHelpers()
+  if (getOption("actel.debug", default = FALSE)) { # nocov start
+    file.copy(paste0(tempdir(), "/temp_log.txt"), paste0(tempdir(), "/latest_actel_error_log.txt"))
+  } else {
+    file.rename(paste0(tempdir(), "/temp_log.txt"), paste0(tempdir(), "/latest_actel_error_log.txt"))
+    deleteHelpers()
+  }
 } # nocov end
 
 #' Recover latest actel crash log
